experimental/multimodal_assistant/requirements.txt
pymupdf==1.23.15
gspread==6.0.0
langchain
langchain_community
langchain_core
langchain_nvidia_ai_endpoints==0.1.0
pandas==2.2.0
Pillow==10.2.0
pydantic==2.5.3
pymilvus==2.3.5
python_pptx==0.6.23
qdrant_client==1.7.1
Requests==2.31.0
streamlit==1.30.0
streamlit_analytics==0.4.1
streamlit_feedback==0.1.3
torch==2.1.2
transformers==4.37.2
